backend/requirements.txt
# FastAPI and Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
python-multipart==0.0.6
python-dotenv==1.0.0

# Strands Agents Framework
strands-agents==1.4.0
strands-agents-tools==1.0.0

# AWS SDK and Tools
boto3==1.34.0
botocore==1.34.0

# Ollama Integration
ollama==0.1.7

# Database and Storage
psycopg2-binary==2.9.9
redis==5.0.1
sqlalchemy==2.0.23
alembic==1.13.1

# Vector Database for RAG
chromadb==0.4.18
sentence-transformers==2.2.2

# Data Processing
pandas==2.1.4
numpy==1.24.3

# HTTP and Web
aiohttp==3.9.1
requests==2.31.0
beautifulsoup4==4.12.2

# Security and Encryption
cryptography==41.0.8
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0

# Utilities
pydantic==2.5.0
asyncio-throttle==1.0.2
python-slugify==8.0.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2